+++ framework/src/main/java/com/bytex/snamp/connector/metrics/NumericGauge.java
@@ -8,7 +8,7 @@ package com.bytex.snamp.connector.metrics;
  */
 interface NumericGauge extends Metric {
     double getDeviation();
-    double getQuantile(final double quantile);
+    double getQuantile(final float quantile);
     double getLastMeanValue(final MetricsInterval interval);
 
     @Override
